renamed interface/data_loader into interface/loader for more readability
